Remove the run-time implementation fallback from certified FIPS builds

+++ src/internal.c
@@ -29699,7 +29699,13 @@ static const char* wolfSSL_ERR_reason_error_string_OpenSSL(unsigned long e)
 }
 #endif /* OPENSSL_EXTRA || OPENSSL_EXTRA_X509_SMALL || HAVE_WEBSERVER || HAVE_MEMCACHED */
 
-wc_static_assert((int)WC_LAST_E <= (int)WOLFSSL_LAST_E);
+/* Compare as long, not int.  MSVC C5287 ("operands are different enum types")
+ * still fires through an (int) cast, because int is the enums' own underlying
+ * type so the conversion is a no-op and the operand keeps its enum type.  A
+ * cast to long is a real conversion and clears it; measured with cl /W4 on
+ * both forms.  Both values are enum constants representable in int, so the
+ * widening changes nothing about what is asserted. */
+wc_static_assert((long)WC_LAST_E <= (long)WOLFSSL_LAST_E);
 
 const char* wolfSSL_ERR_reason_error_string(unsigned long e)
 {
